+++ src/main/java/com/hackathonteam1/refreshrator/controller/AuthController.java
@@ -3,9 +3,9 @@ package com.hackathonteam1.refreshrator.controller;
 import com.hackathonteam1.refreshrator.authentication.AuthenticationExtractor;
 import com.hackathonteam1.refreshrator.authentication.JwtEncoder;
 import com.hackathonteam1.refreshrator.dto.ResponseDto;
-import com.hackathonteam1.refreshrator.dto.request.LoginDto;
-import com.hackathonteam1.refreshrator.dto.request.SigninDto;
-import com.hackathonteam1.refreshrator.dto.response.TokenResponseDto;
+import com.hackathonteam1.refreshrator.dto.request.auth.LoginDto;
+import com.hackathonteam1.refreshrator.dto.request.auth.SigninDto;
+import com.hackathonteam1.refreshrator.dto.response.auth.TokenResponseDto;
 import com.hackathonteam1.refreshrator.service.AuthService;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
